feat(financeiro): trava saldo, débito automático de fixas e UX de contas
Adiciona trilha append-only de movimentação, lock pessimista no saldo,
reparo idempotente com flag, alertas operacionais e rede de regressão P0.
Implementa débito automático de obrigações fixas por conta no vencimento.
Remove recalcular saldo confuso e permite editar saldo direto no modal de contas.

+++ backend/src/main/java/com/consumoesperto/service/CartaoCreditoService.java
@@ -567,10 +567,8 @@ public class CartaoCreditoService {
      */
     public BigDecimal getTotalLimiteCredito(Long usuarioId) {
         // Busca o total do limite de crédito de todos os cartões ativos do usuário
-        Double total = cartaoCreditoRepository.getTotalCreditLimitByUsuarioId(usuarioId);
-        
-        // Converte para BigDecimal ou retorna zero se não houver cartões
-        return total != null ? BigDecimal.valueOf(total) : BigDecimal.ZERO;
+        BigDecimal total = cartaoCreditoRepository.getTotalCreditLimitByUsuarioId(usuarioId);
+        return total != null ? total : BigDecimal.ZERO;
     }
 
     /**
